move json encoder and transformer to utils
