+++ src/lib/decorators/thread/ThreadDecorator.test.ts
@@ -1,4 +1,4 @@
-import { JsonTransformer } from '../../JsonTransformer';
+import { JsonTransformer } from '../../utils/JsonTransformer';
 import { ThreadDecorator } from './ThreadDecorator';
 
 describe('Decorators | ThreadDecorator', () => {
