+++ src/lib/decorators/timing/TimingDecorator.test.ts
@@ -1,4 +1,4 @@
-import { JsonTransformer } from '../../JsonTransformer';
+import { JsonTransformer } from '../../utils/JsonTransformer';
 import { TimingDecorator } from './TimingDecorator';
 
 describe('Decorators | TimingDecorator', () => {
